demo_loader/requirements.txt
pymongo 
openai 
python-dotenv
tiktoken
langchain-openai
azure-storage-blob
azure-identity
langchain==0.1.6
langchain-community==0.0.19
langchain-core==0.1.23